requirements/base.txt
Django==5.0.8
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
Pillow==10.4.0
pandas==2.2.2
numpy==2.1.0
scikit-learn==1.5.1
xgboost==2.1.1
joblib==1.4.2
faker==28.4.1
python-dateutil==2.9.0
